Move Preprocessor parsing-related methods to Parsers
They're only used from there.
